Make AST generic over string values
This commit introduces a Text trait that is
used for representing text data in the AST.
This enables using the ast with &str references only,
with Cow, or just as now in an owned fashion (with String).

To prevent a wrapper type, the Text trait has an associated type 'Value'
that represents the actual value and has to be used in the AST nodes.