[Federation] Add error codes to snapshot results (apollographql/apollo-server#2785)
Introduce error codes to snapshot results for validations in order
to ensure their stability.

Created a new serializer for GraphQLErrors in order to display the
error codes we attach to GraphQLError.extensions.code. This
serializer should be used wherever we're displaying GraphQLErrors
that we create.
Apollo-Orig-Commit-AS: apollographql/apollo-server@8ad52d428773896df11cf32887258a6ca7731e6f