fix: @provides validation for object types (apollographql/apollo-server#2901)
The previous post-composition validators `providesFieldsMissingExternal` and
`providesFieldsSelectInvalidType` conflicted with the [federation spec][spec].

In the case of "missing external" validations, the previous code looked for
the provided field on the wrong type. Using the following example:

```gql
type A {
  b: C @provides(fields: "d")
}

type C {
  d: Int
}
```

The validator was looking for field `d` on type `A`. It now looks at the type
of `b` to correctly lookup field `d` on type `C`.

Similarly, the "fields select invalid type" validator looked up the provided
field incorrectly. The fix is basically the same in both validators.

[spec]:https://www.apollographql.com/docs/apollo-server/federation/advanced-features/#using-denormalized-data
Apollo-Orig-Commit-AS: apollographql/apollo-server@183ed5dca8c2ed07bb052ebe99e6a008b290d5ea