Convert trace failure into metric
Previously, the logic within executing a query plan would throw an error
and fail the entire operation if it could not parse a trace. Instead, it
adds the information that it failed to parse the trace into the overall
fetch node itself. This commit also adds some comments.

Apollo-Orig-Commit-AS: apollographql/apollo-server@217fab3fa71896086822846ae3496fc7a8a999af