gateway: proxy downstream service errors (apollographql/apollo-server#3019)
* gateway: proxy downstream service errors

In the gateway, we currently squash all errors from a single service
into a single `Error` with a generic message, such as `Error while
fetching subquery from service "registry"`. In other words, the gateway
globally overwrites errors from the underlying services and the original
errors are appended to `Error.extensions.downstreamErrors`.

https://github.com/apollographql/apollo-server/blob/e849670ffe9e66ad05099aa6b9e9daeea055d489/packages/apollo-gateway/src/executeQueryPlan.ts#L242

This changes to behavior, so that the errors are proxied in their same
form rather than nested into a single error. Ultimately the underlying
services are better equipped than the gateway to expose/mask their own
errors with `formatErrors` and this removes the gateway as the
bottleneck of configuration. The original desire with the message mask
was to retain service info / context. To accomplish this, we add the
`serviceName` to the `Error.extensions` object for each. so users can
investigate the results.

* gateway: update changelog

Apollo-Orig-Commit-AS: apollographql/apollo-server@dc2af9c70c9ac78f2a488b0a1b01f5aafba3ceec