Fix value type behavior within composition and execution (apollographql/apollo-server#3182)
This commit implements value types as a core part of composition
and query planning. With this work, we now:

* build another metadata map for tracking value types
* attach value type metadata to schema type nodes
* attach value type metadata to fields on value types
* make query planning decisions based on value type metadata

With value types, it doesn't make sense for a single service to own
the fields on said value type. Update the types and implementation to reflect
this.
Apollo-Orig-Commit-AS: apollographql/apollo-server@1852033fba9a074558bd9801707fef3da8f75c88