types: Adjust prior assumptions for new graphql 14.5.x types. (apollographql/apollo-server#3225)
Fix incorrect assumption that the `ExecutionResult` would never be `null`.  This is not true, and is now enforced in the typings for graphql@^14.5.4.

Additionally, while `GraphQLInputField` used to be assignable to `GraphQLArgument`, it no longer is since `description` is required on a `GraphQLArgument` while optional on `GraphQLInputField`.
Apollo-Orig-Commit-AS: apollographql/apollo-server@befe9d17909cb7d9691288a2a2f56f0b0ba833e8