Add `context` object to `GraphQLDataSource.didReceiveResponse` arguments (apollographql/apollo-server#3360)
This adds a third parameter to the `didReceiveResponse` method on
a GraphQLDataSource. This gives the hook access to the context
object, allowing it to read and modify the context as necessary.

As demonstrated by the included test, this enables a GraphQLDataSource
to inspect the response and add information to the context accordingly.

This also provides parity with another hook, willSendRequest, which already
has access to the context object.
Apollo-Orig-Commit-AS: apollographql/apollo-server@8881d45dfbc6409b27c0ea0fcc30c8b2ff2caebb