Gateway cached DataSource bug (apollographql/apollo-server#3412)
This fixes a bug in the gateway that prevents it from constructing
new `DataSource`s when composition updates and the url of an
underlying service changes. Currently the creation of a `DataSource`
is cached based on its `serviceName`, but really the caching should
depend on both the `serviceName` and `url` of the `ServiceDefinition`.
Apollo-Orig-Commit-AS: apollographql/apollo-server@21a7853ce43733afcecb14c9debd491147243154