Begin supporting executable directives in federation (apollographql/apollo-server#3464)
This commit is the very start of executable directives in federation.
It removes extra filtering we currently do to allow any directives that
were part of the service definitions to be maintained in the final
schema. The final implementation is a model where if an executable
directive is present, *all* services must both contain and exactly match
the ones found in other services. This is because executable directives
can't be limited to *specific* types/fields, but must be usable on any
location they are defined. Long term we may be able to move support for
common executable directives to the gateway but custom ones will always
require them to be present in all services (even as a no-op)

* Update post-composition validator API

Validators may need knowledge of the whole serviceList in order to
perform certain validations. This change extends the API to call
post-composition validators with this additional information.

* Update current validators
* Update tests
* Create PostCompositionValidator type

* Add directive metadata to composed schema

This commit appends all of the DirectiveDefinitionNodes to the
composed schema (at their respective GraphQLDirective) to be
inspected by validators. These definitions are indexed by the
service that they belong to in order to provide useful validation
messages.

* Add validator for permitting only custom ExecutableDirectives

* Add validator for the enforcing the presence of custom directives within EVERY service

* Add validator for enforcing custom directive identicality across services

* Rename DefinitionsMap and ExtensionsMap

* Add integration tests for custom directives

These tests verify that the query plan includes custom
directives in requests to underlying services.

They also ensure that validation errors are as expected
when services define custom directives in an invalid way.
Apollo-Orig-Commit-AS: apollographql/apollo-server@e3b118b2945f862a473a10324889b43435513c9f