Adjust deceiving error message about previous schemas and falling back.
This message was being issued when a new server starts up, prior to ever having
a schema, when a storage secret (or any other artifact) can't be fetched
from GCS (or any error within `updateServiceDefinitions`) despite the fact
that there may be no `previousSchema`.

While I could use `previousSchema` to change the error message, I don't
think that it's this methods concern to decide what happens in the event of
an error.  I think it's only this methods concern to actually do the update,
if it is in fact successful.

Apollo-Orig-Commit-AS: apollographql/apollo-server@06c3267d27f72d2cbf071bdbfc98886a0250fb06