Adjust schema rollback test for less flakiness
A long `await`ed change (hah), this should put a nail in the schema
rollback test flakiness once and for all. This solution is a
code-driven approach rather than a time-driven approach. It depends
directly on the schema change listener functionality already built-in
to the gateway.

cc @abernix

Apollo-Orig-Commit-AS: apollographql/apollo-server@8996158d5b2e3559c3316bfe64532f569e935ad9