Fix Typescript generic typing for datasource contexts (apollographql/apollo-server#3865)
This is a fix for the `TContext` typings of the gateway's exposed
`GraphQLDataSource` implementations. In their previous form, they
didn't work as intended, or in any manner that's useful for typing the
`context` property throughout the class methods. This introduces a type
argument `TContext` to the class itself (which defaults to
`Record<string, any>` for existing implementations) and removes the
non-operational type arguments on the class methods themselves.
Apollo-Orig-Commit-AS: apollographql/apollo-server@5afdb4e9e5083b76094cd33015a9ee16762d68eb