Gateway downstream health checks (apollographql/apollo-server#3930)
This commit introduces a new configuration option to the gateway: performServiceHealthChecks. This option adds an additional
precautionary step to the gateway load and schema update code paths.

A gateway with this setting turned on will send a simple graphql query
to all downstream services ({ __typename }) to ensure the service is
live and responsive. It performs these queries on initial schema load
as well as during a schema update. The interesting failure behaviors are:

* On load: throw an error, failure to load.
* On schema update: "rollback" (never roll forward). Log the event, but continue serving requests with the old schema.
Apollo-Orig-Commit-AS: apollographql/apollo-server@5eef2a663c82dbe5f450806a6429791db117a4e5