Convert all line endings into single \n
Windows filesystems can read line endings as \r\n even if they are
authored in a *nix filesystem. The windows test suite was previously
tfailing because fixtures of the tests were being split on a \n---\n
where the windows CI was seeing \r\n---\r\n. This is coarse grain fix to
convert all fixtures to a single \n line ending.