clarity: Change `getRemoteConfig` to `getManagedConfig` internally.
This is a "remote" configuration, but specifically, it's the "stored"
configuration which lives within Apollo Graph Manager's storage facilities.

This is not to be confused with the "remote" config which can be fetched
from a remote endpoint that is running a federated service which is typical
in `serviceList` mode.

Apollo-Orig-Commit-AS: apollographql/apollo-server@84fdcafd0b9024cf64ae88fedd52ec7adaf8c11f