Configure error reporting for CDN worker
This commit creates a new file for the worker that sets up reporting
errors to sentry.io. It also supports this integration with a deployment
addition in GitHub actions for the CDN. The final additions is that of
an index.html which came from inital error reporting when loading the
worker locally.