Rewrite the Readme and Contributing guide to be more welcoming
This is a first pass but I'm pretty happy with where it landed. It
should be much easier for people to get started on the project with
these guides. I'd love for us to write another on "how to add a command"
when we have established a practice that we like for that