Test the CLI installer shell script
Since the CLI installer is the second point of failure after the CDN
proxy, it is pretty important that we test this installer to prevent
accidental issues breaking everyones install path (especially for CI
environments).

This commit sets up the awesome [bats](https://github.com/bats-core)
testing framework for shell scripts. It includes coverage for the
current installer plus an attempt at coverage support for the project
all together. NOTE, codecov remapping of files isn't working so coverage
values will be there, but looking at the report online doesn't seem to
be working for the specific installer script.

The test framework is installed by npm to keep the "get up and running"
pretty easy still for the CDN