feat(gateway): relocate downstream query construction (apollographql/apollo-server#4018)
This commit moves query construction / printing into the
buildQueryPlan fn (previously in executeQueryPlan).

This decision was informed by @abernix, who put together a
wonderful benchmark of the gateway. We noticed that the call
to graphql-js's `print` fn was incredibly expensive for large
queries. To offset the cost of this, this diff now hoists all
of this work into `buildQueryPlan`, where the result will
ultimately be cached.
Apollo-Orig-Commit-AS: apollographql/apollo-server@201630ad284754248fc9ab6ebedc7506fcc3d951