Rework test corpus into more, smaller tests, and generate them at build time.
We now have a build.rs which generates tests.rs from a listing of .graphql files
in the tests dir. I did this because I wanted to use include_str! to include the
example and optional canonical form at build time and break up the corpus into
more, smaller tests. Previously, we were asserting multiple snapshots in the same
test, which led to waterfalling—if an early snapshot failed, future ones in the
same test wouldn't be checked or generated. You'd have to run the test suite
multiple times to ensure you'd generated them all, which was error prone.

This also simplifies the process of adding tests: you just add a .graphql sample,
and maybe a .canonical.graphql file, if the sample isn't in canonical form.