Fix IntelliJ Configuration
After trying to work with the previous configuration I set up in https://github.com/apollographql/apollo-cli/pull/40, I noticed some annoyances, specifically that the parent module created by the Rust client was set up to be a Java module, but the Rust plugin ignores that.

Not an issue if we had only Rust, but then there's the `cdn` module too, which was detected to be Java and is not a Rust module, so IJ was complaining that there's no Java SDK even though we don't need one.

This is an attempt to fix that. Still not great, WIP.