IntelliIJ configuration (attempt 2) (#43)
Trying to work on the repo with the previous configuration (from #40) worked well for coding, code completion and navigation, however, when trying to build the project, IntelliJ complained there is no Java SDK defined. 

I did some digging and found 2 things:
+ The set up in this PR is slightly better, in that IntelliJ doesn't complain anymore and we can have a module for each project (`cli`, `graphql-parser`, `cdn`).
+ There's a bug with the Rust plugin that I reported here (https://github.com/intellij-rust/intellij-rust/issues/5424) that causes the cli to have this double nested module. Apart from the UI looking like that, I couldn't see any other issues. 
+ When following the instructions from the Rust plugin site for how to set up the IntelliJ project, we end up with the wrong configuration, that is also reported in the same bug. 

Hopefully that bug will be addressed, at which point we may or may not have to change this again. I'll stay on top of that.

Since no one else uses IJ on the team just yet, I'll just go ahead and merge.