Support updating the CLI from the CLI
This commit supports a dedicated update command for the CLI.
It also lands a background task on each run that checks for an update.
We may want to slow the task to run only once a day, but right now
I don't see the harm in an async attempt to check when running commands.

I also cleaned up the setup command here to be less agressive on
writing to shell configs both in the installer and in the setup command
itself.