Rework the CDN to scale and support telemetry
This commit cleans up the structure of the worker so it can scale better
from a code organization point of view. It also adds support for
telemetry reporting which was the reason for the clean up.

Currently telemetry goes directly to segment and the envs have already
been setup with the segment key information.