Only check for updates once per day
This commit adds support to checking the update status of the CLI at
most once per day. We do this by writing a $HOME/.apollo/version.toml
file to the system with the latest version and a timestamp. This should
prevent us from overchecking or reporting when we don't need to do that
work.