chore(deps): update dependency graphql to v15
In a general sense this just updates tests to accommodate the new error
conditions and schema hashes since a lot of the work has already been done
in apollographql/apollo-server#3712 / 9db3db84.

A larger note is due to explain the updating the snapshots.  Those changes
are due the fact that types returned by, e.g.  `getImplementation`, will now
be dependent on the order that those types first appear in the schema.

See the referenced issues, the first of which implemented the change and the
second which indicates the motivation.

Ref: https://github.com/graphql/graphql-js/pull/2410
Ref: https://github.com/graphql/graphql-js/pull/2362
Ref: apollographql/apollo-server#3712 (9db3db84428736079605edd47c2ee3cf516)

Apollo-Orig-Commit-AS: apollographql/apollo-server@e009f527868b68e9336a58953ba659de48b28f6f