Model for QueryPlan in Rust (#60)
This PR implements model objects for a `QueryPlan`, pretty much ported from how we do it in JS.

The goal for the `QueryPlan` is to
1. Contain owned data so that they can be cached
2. Contain only the data required to execute a QueryPlan

This model assumes that we will build a `QueryPlan` in a way that's similar to how it's built in JS, where we calculate and store the `operation` that we'd need to invoke in each `FetchNode. Note that in JS we call this field `source`, but `operation` is a more appropriate name.