Fix query plan missing fields due to loss of specific scope
The query planner associate to each occurrence of a field in a query a
"scope". That scope records a so-called "parent type" as well as all the
types the object could have when reaching that field.

For instance, with:
```graphql
{
  ... on X {
    ... on Z {
      f
    }
  }
  ... on Y {
    ... on Z {
      f
    }
  }
}
```
the planner associate `Z` as parent type of both occurrences of `f` (
because it's mot the "immediate" condition applied to it in both
cases), but also record that the 1st occurrence is necessary a `X`
while it is necessary a `Y` in the 2nd occurrence.

When handling fields a particular level, the planner groups fields by
field name and parent type. In the case above, it treats both
occurrences of `f` together, because they have the same parent type `Z`.

However, when treating that group, the planner incorrectly ignores the
fact that while the parent types are the same, the scope are not fully
similar, and it only handles the 1st occurrence of the field. As
a result, the generated plan ends up only fetching the first
occurrence, which is incorrect (concretely, the gateway never return `Y`
objects while it should).

This commit fixes the problem by changing the way fields are grouped.
Instead of being grouped just by parent type, they are grouped by
their scope. So, in the example above, both occurrences of `f` are
treated separately.

Fixes: #396