JSON De/Serialization for `QueryPlan` (compatible with JS)
This PR adds the ability to serialize and deserialize a `QueryPlan` into JSON in the same format as it is on the JS side.
We will use then be able to use the same cucumber tests that we have in apollo-gateway to test the Rust implementation of the query planner.

The PR heavily uses the `serde` and `serde_json` crates, and their provided annotations to to achieve the format compatibility.