JSON De/Serialization for `QueryPlan` (compatible with JS) (#88)
This PR adds the ability to serialize and deserialize a `QueryPlan` into JSON in the same format as it is on the JS side.
We will use then be able to use the same cucumber tests that we have in apollo-gateway to test the Rust implementation of the query planner.

The PR heavily uses the `serde` and `serde_json` crates, and their provided annotations to to achieve the format compatibility. I had to do some minor changes to the model to support this, nothing that really affects ergonomics (for context, the model code was implemented very recently and when I wrote it then I could have just as easily decided to model it like in this PR, but only happened to not do that).

The only real workaround that I had to add was to create a private alternative, single variant enum (`QueryPlanSerde`) for QueryPlan that will serialize with a `kind` field.