Visitors and other utility traits (#83)
Co-authored-by: Ran Magen <ran@apollographql.com>
Co-authored-by: Ashi Krishnan <queerviolet@github.com>

This PR adds the visitor pattern to the query and schema ASTs. It also adds a `Name` trait, which AST nodes implement if they have a `name`. Other similarly-useful traits will be forthcoming.

In addition to being able to visit nodes, we're also adding the ability to map over the AST tree. Tree maps are like visitors, only they return an arbitrary output type.