Change `visit::Map`'s `merge` to take ownership of `child` (#107)
If `Output` is a container type (e.g. a `Vec<T>`), then `merge`'s ergonomics becomes almost useless.
The _borrowed_ child means that merge cannot take ownership of the contents of the container, and so any attempt to implement a mapping that returns a container is rendered impossible without copying/cloning the contents of the container.

## Why is this an ok change? Don't we need to maintain ownership?

Turns out, no. :)

The function that calls `merge` is `Fold.pop`; There is an implicit invariant on the `Fold.pop` function, where if we push into the stack, we definitely call `Fold.pop` again, and override the owned value.
Currently, the owner of the value passed as `child` is `self.output`, but the only call to `merge` is used to `push` to the stack, which means `Fold.pop` will be called again, and the first line of `Fold.pop` overrides `self.output`, and so `self.output` currently maintains ownership for a value that is never read.

In the scenario where we don't call `merge` (the `stack` is empty after popping), the invariant is that this is the scenario where we finished walking the tree, and `output` may be read.

This PR makes that clearer: we `pop` into a local `output` variable, and if we're done visiting, we give ownership to `self.output`, and if we're still pushing into the stack, we give ownership to the `merge` function.

Previously, in case the stack was empty when `Fold.pop` was called, `None` was assigned to `self.output`, I'm not sure when that can happen, but that invariant has not been changed in case it was relied upon.

## Unrelated
those "hidden" lines in doctests (prefixed with `#`) are making examples more confusing and are not needed as far as I can tell.
Doctests run their code in a `main`, so adding an additional `main` only to call `parse().unwrap()` seems more confusing for an example, that contains `?` usages but will not display the `Result` return type, and so using the example as documented will not be as clear.