query-planner-wasm (#127)
Add a crate that exposes the high level QueryPlanner API via Wasm.

## API and `unsafe`

We ran some benchmarks and discovered that parsing the schema is ~50% of the work when query planning. It therefore became a priority to make sure we only parse the schema once.

I had to figure out a way to keep some memory alive on the Wasm side, so that the JS side could interact with it. There is a compiler restriction that doesn't allow us to pass a value with lifetimes (in our case `QueryPlanner<'q>`) back to the JS side.

Theoretically though, we really just need a pointer to it. In Rust, converting a reference to a raw pointer is safe, but using that pointer is not safe.

The API implementation uses `unsafe` in a way that we think is safe, with some limitations. We instantiate the `QueryPlanner`, store it statically, and return a pointer to it. A later function to plan a query sends in the pointer to be used, and we de-reference that pointer.

As the code is now, we _could_ actually not at all return a pointer, and when we need to plan a query, we could assume that there is a planner available statically and use that one, however, we believe that very soon, we'd need to support a transitional state when the schema changes in the gateway, but there are in flight request that still need to plan.

## Limitations

This implementation assumes there is only one active schema for which we're planning queries. This is not actually the reality. The gateway can hot-swap its running schema, and theoretically be in a transitional state where it has in flight requests that have been planned with schema A and validated with schema B. We need to explicitly write code to ensure everything that was planned with A uses A until the request is complete. When we figure out the JS side, we'll update this code to enable multiple planners and explicitly destroying an unused planner.

## Naming 

We decided to have `-wasm` crates for each Rust "twin" crate instead of a one monolithic "bridge" crate that would expose multiple Rust crates' implementations via wasm. 
In order for things to show up nicely on the filesystem, we're naming these crates as `X-wasm` where `X` is the name of the Rust crate we're exposing.

## Integration

The branch on the `apollo-server` side to integrate this is https://github.com/apollographql/apollo-server/tree/trevor/integrate-wasm-query-planner