Initial working implementation of execution
This commit is a first take at porting over execution from the
TypeScript implementation as well as surrounding scaffolding for a PoC
server runtime and process manager. This is an initial stab for feedback
and will be split up into multiple commits for review in different PRs
around structure, execution, and basic server scaffold after testing is
done.