Use TypeScript-specific `.vscode/settings.json` attrs only on `TypeScript`
Prior to this change, these settings were overzealous in that they applied
to the all code in this repository (not just JavaScript and TypeScript),
which is out of alignment with the formatting standards used by other
language ecosystems (e.g., Rust).

This was inadvertently introduced this way.

Also, make it so that highlighting isn't broken on GitHub for this file
which VSCode interprets as JSON5, despite the `.json` extension.