Setup CircleCI/Lerna/npm to compile `query-planner-wasm` prior to building.
This allows us to compile and depend upon the local version of the WASM
query planner package and get the tarballs that are packed with the right
version.

This relieves the `Cargo.toml` from its responsibility of managing the
`version`, and instead passes that responsibility to `lerna` which allows
Lerna to do that with full accountability for the `dependencies` in this
monorepo that depend on the WASM package itself to be properly published.